misc/HtmlTestSuite/results/test-0407.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="mstensho@chromium.org" rel="author" title="Morten Stenshorne"/>
<link href="https://drafts.csswg.org/css-scroll-anchoring/" rel="help"/>

</head><body><p>There should be no red below.</p>
<div id="container" style="writing-mode:vertical-rl; overflow:auto; width:300px; height:300px;">
  <div style="width:300px; background:red;"/>
  <div style="width:400px; font-size:16px; line-height:25px;">
    <span id="displayMe" style="color:red; display:none;">
      FAIL<br/>FAIL<br/>FAIL<br/>FAIL<br/>
    </span>
    line<br/>
  </div>
  <div id="displayMeToo" style="display:none; width:300px; background:red;"/>
</div>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
  test(()=&gt; {
      var container = document.getElementById("container");
      var displayMe = document.getElementById("displayMe");
      var displayMeToo = document.getElementById("displayMeToo");
      // Scroll the text container into view.
      container.scrollLeft = -300;
      displayMe.style.display = "inline";
      displayMeToo.style.display = "block";
      assert_equals(container.scrollLeft, -400);
  }, "Line at edge of scrollport shouldn't jump visually when content is inserted before");
</script>
</body></html>
